Terraform-Architect-FullStack/backend/requirements.txt

fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-chroma
langchain-huggingface
sentence-transformers
pydantic
sqlalchemy
passlib[argon2]
python-jose[cryptography]
python-multipart
python-hcl2
diagrams
requests
tqdm
datasets
